Use /usr/bin/env for shebang line, instead of /usr/bin/python
This is changed for all python scripts except diffinfo.
This could not be used with diffinfo, because of the use of
an argument python.  Linux can only pass one argument
to the interpreter in the shebang line.

Thus #!/usr/bin/env python -u results in '/usr/bin/env "python -u"'
which doesn't work.